Backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-auditlog==3.1.2
django-cors-headers==4.7.0
django-filter==25.1
django-jsonfield==1.4.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
openpyxl==3.1.5
pillow==11.2.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
